requirements.txt
pypdf
numpy
sentence-transformers
mcp
grpcio>=1.60
grpcio-tools>=1.60
protobuf>=4.25